fix(iphone): consuming first letter & refactor the useSearchParam
